commit init
This Python code uses the "ipaddress" and "subprocess" libraries to perform a ping test on all possible IP addresses in a given network and print a message if the device responds to the ping.

The code begins by importing the "ipaddress" and "subprocess" libraries. It then defines the network information, including the network address, subnet mask, and default gateway.

Next, the subnet mask is converted to the correct format using the "IPv4Network" class from the "ipaddress" library.

The code uses a "for" loop to iterate over all possible IP addresses in the network. For each IP address, it sends a ping command using the "subprocess.Popen" method from the "subprocess" library with the necessary arguments to send a single ping and wait for a maximum of 500 milliseconds for a response. The result is stored in a variable called "response".

The code checks if the device responded to the ping by checking if the string "Reply" is present in the output of the ping command converted to a string. If the device responds to the ping, the IP address is printed to the screen with the message "is active".