Rename whois_domain to whois_domain.py
In this version, the user will be prompted to input the domain name. Then, the code will check if the domain name starts with "www.". If not, it will add "www." to the beginning of the domain name before passing it to the whois module for domain information retrieval. Then, the domain information will be printed to the console.