Update Tropico_4_Download.html
This is a batch script that performs a series of commands related to network configuration on Windows. It executes the following commands sequentially:   These commands are specific to Windows and can be used to diagnose and reconfigure network connection in a Windows environment.

Usage instructions:    Clone the repository to your local environment.
Run the comando-ipconfig.bat file by double-clicking on it or typing its name in the Windows command prompt.
The commands will be executed sequentially, with a 20-second interval between each command.
Check the output of the commands in the command prompt to verify the results.
Please note that executing network commands can have consequences and requires proper knowledge of networking and system configuration. Use this script with caution and at your own risk.  Downloaded user with caution.